feat: Simplify issue form.
