--- src/components/ADempiere/Form/Issues/index.vue
@@ -1,243 +1,0 @@
-<!--
- ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
- Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
- Contributor(s): Elsio Sanchez Elsiosanches@gmail.com https://github.com/Elsiosanchez
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program. If not, see <https:www.gnu.org/licenses/>.
--->
-
-<template>
-  <el-container :style="isNewIssues ? 'height: 90vh;padding: 0px;display: contents' : 'height: 90vh;padding: 0px;'">
-    <el-main style="padding: 0px;overflow: auto;">
-      <list-issues v-if="!isNewIssues" />
-    </el-main>
-    <el-footer v-if="isNewIssues" height="20%">
-      <comment
-        :table-name="tableName"
-        :record-id="recordId"
-      />
-    </el-footer>
-  </el-container>
-</template>
-
-<script>
-import {
-  defineComponent, computed, ref
-} from '@vue/composition-api'
-
-import store from '@/store'
-
-// Components and Mixins
-import Comment from './component/Comment.vue'
-import RecordTime from './recordTime.vue'
-import ListIssues from './ListIssues'
-
-// Constants
-import { REQUEST_WINDOW_UUID } from '@/utils/ADempiere/dictionary/form/Issues.js'
-
-// Utils and Helper Methods
-import { formatDate } from '@/utils/ADempiere/formatValue/dateFormat'
-import { zoomIn } from '@/utils/ADempiere/coreUtils.js'
-import { getImagePath } from '@/utils/ADempiere/resource.js'
-
-export default defineComponent({
-  name: 'Issues',
-
-  components: {
-    // Editor
-    Comment,
-    ListIssues,
-    RecordTime
-  },
-
-  props: {
-    tableName: {
-      type: String,
-      required: false
-    },
-    recordId: {
-      type: Number,
-      required: false
-    }
-  },
-
-  setup(props) {
-    const message = ref('')
-    const filter = ref('')
-    const priority = ref('')
-    const typeRequest = ref('')
-
-    const listIssues = computed(() => {
-      return store.getters.getListIssues
-    })
-
-    const isNewIssues = computed({
-      // getter
-      get() {
-        return store.getters.getNewIssues
-      },
-      // setter
-      set(newValue) {
-        store.commit('setNewIssues', newValue)
-      }
-    })
-
-    const currentIssues = computed(() => {
-      return store.getters.getCurrentIssues
-    })
-
-    const isShowTitleForm = computed(() => {
-      return store.getters.getIsShowTitleForm
-    })
-
-    const styleAllRequestBoxCard = computed(() => {
-      if (isShowTitleForm.value) return 'height: 80%;overflow: auto;'
-      return 'height: 90%;overflow: auto;'
-    })
-
-    function dueTypeColor(issue) {
-      const { due_type } = issue
-      const { value } = due_type
-      let color = '#3fb950'
-      if (value === '5') {
-        color = 'orange'
-      } else if (value === '3') {
-        color = '#ff2121'
-      }
-      return color
-    }
-
-    function selectIssue(issue) {
-      isNewIssues.value = !isNewIssues.value
-      store.dispatch('changeCurrentIssues', issue)
-    }
-
-    function newIssues(issue) {
-      isNewIssues.value = !isNewIssues.value
-      store.dispatch('changeCurrentIssues', issue)
-    }
-
-    function loadIssues() {
-      store.dispatch('listRequest', {})
-      store.dispatch('findListMailTemplates')
-    }
-
-    function zoomIssues(issues) {
-      zoomIn({
-        uuid: REQUEST_WINDOW_UUID,
-        params: {
-          filters: [
-            {
-              columnName: 'UUID',
-              value: issues.uuid
-            }
-          ]
-        }
-      })
-    }
-
-    function avatarResize(user) {
-      const { avatar } = user
-      const { uri } = getImagePath({
-        file: avatar,
-        width: 20,
-        height: 20,
-        operation: 'resize'
-      })
-      return uri
-    }
-
-    loadIssues()
-
-    return {
-      message,
-      listIssues,
-      //
-      priority,
-      typeRequest,
-      filter,
-      isNewIssues,
-      currentIssues,
-      styleAllRequestBoxCard,
-      isShowTitleForm,
-      // methods
-      dueTypeColor,
-      formatDate,
-      avatarResize,
-      selectIssue,
-      newIssues,
-      loadIssues,
-      zoomIssues
-    }
-  }
-})
-</script>
-
-<style lang="scss">
-.all-request-box-card {
-  padding: 0px;
-  .el-card__body {
-    padding: 0px;
-  }
-}
-.table-list-request {
-  .el-table td.el-table__cell div {
-    display: flex;
-    padding: 0px;
-  }
-  .el-table thead {
-    display: none;
-  }
-  .el-table--medium .el-table__cell {
-    padding: 0px;
-  }
-}
-.p {
-  margin: 0px !important;
-}
-.el-card__body {
-  padding: 20px;
-  padding-top: 0px !important;
-  padding-right: 20px;
-  padding-bottom: 0px!important;
-  padding-left: 20px;
-}
-.tui-editor .te-preview-style-vertical .te-preview {
-  float: left;
-  width: auto !important;
-}
-.te-preview {
-    overflow: auto;
-    width: auto !important;
-    padding: 0 25px;
-    height: 100%;
-}
-
-.issues-card {
-  .el-card__body {
-    padding-top: 0px !important;
-    padding-right: 0px;
-    padding-bottom: 0px!important;
-    padding-left: 0px;
-  }
-}
-// .el-descriptions-item__label:not(.is-bordered-label) {
-//   padding-top: 10px;
-// }
-</style>
-<style scoped>
-.scroll-chats {
-  width: 100%;
-  height: 500px;
-}
-</style>
